requirements.txt
alembic==1.4.3
attrs==20.3.0
click==7.1.2
coverage==5.3
Flask==1.1.2
Flask-Admin==1.5.7
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.7
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3
